@@ -1,57 +1,63 @@
-require 'bundler/capistrano'
-require 'capistrano-rbenv'
+# config valid only for current version of Capistrano
+lock '3.8.1'
 
-set :application, "team_registration"
-#set :repository, "ssh://petr@kacr.info:10522/home/git/repositories/team_registration.git"
-set :repository, 'git@bitbucket.org:petrpalata/team_registration.git'
-set :branch, "master"
-#set :deploy_via, :remote_cache
-set :scm, :git
-set :bundle_flags,    "--deployment"
-set :bundle_without,  []
-set :use_sudo, false
+set :application, 'team_registration'
+set :repo_url, 'git@bitbucket.org:petrpalata/team_registration.git'
 
-set :rbenv_ruby_version, "1.9.3-p551"
+# Default branch is :master
+# ask :branch, `git rev-parse --abbrev-ref HEAD`.chomp
 
-load 'deploy/assets'
-
-default_run_options[:pty] = true
-
+# Default deploy_to directory is /var/www/my_app_name
 set :deploy_to, '/home/kwertan/team_registration'
 
-set :ssh_options, { :forward_agent => true }
+# Default value for :scm is :git
+# set :scm, :git
 
-role :assets, "agility2017.cz"
-role :app, "agility2017.cz"
-role :web, "agility2017.cz"
-role :db,  "agility2017.cz", :primary => true
+# Default value for :format is :pretty
+# set :format, :pretty
 
-set :user, "kwertan"
+# Default value for :log_level is :debug
+# set :log_level, :debug
 
-set :stages, ["production"]
-set :default_stage, "production"
-
-set :repository_cache, "git_cache"
-set :deploy_via, :remote_cache
+# Default value for :pty is false
+# set :pty, true
 
 # Default value for :linked_files is []
 set :linked_files, fetch(:linked_files, []).push('config/database.yml', 'config/secrets.yml')
 
 # Default value for linked_dirs is []
-set :linked_dirs, fetch(:linked_dirs, []).push('log', 'tmp/pids', 'tmp/cache', 'tmp/sockets', 'vendor/bundle')
+set :linked_dirs, fetch(:linked_dirs, []).push('log', 'tmp/pids', 'tmp/cache', 'tmp/sockets', 'vendor/bundle', 'public/system')
 
-namespace :passenger do
-    desc "Restart Application"
-    task :restart, :roles => :app, :except => { :no_release => true } do
-        run "touch #{current_path}/tmp/restart.txt"
-    end
-end
+# Default value for default_env is {}
+# set :default_env, { path: "/opt/ruby/bin:$PATH" }
+
+# Default value for keep_releases is 5
+# set :keep_releases, 5
+# user which will execute commands
+
+set :user, "kwertan"
+set :stages, ["production"]
+set :default_stage, "production"
+
+# allow ssh agent forwarding to local host
+set :ssh_options, {
+    forward_agent: true
+}
+
+set :rbenv_type, :user
+set :rbenv_ruby, File.read('.ruby-version').strip
+set :rbenv_map_bins, %w{rake gem bundle ruby rails}
+set :rbenv_roles, :all # default value
 
 namespace :deploy do
-    [:start, :stop, :restart].each do |t|
-        desc "#{t} task is a no-op with mod_rails"
-        task t, :roles => :app do ; end
-    end
-end
 
-after :deploy, "passenger:restart"
+  after :restart, :clear_cache do
+    on roles(:web), in: :groups, limit: 3, wait: 10 do
+      # Here we can do anything such as:
+      # within release_path do
+      #   execute :rake, 'cache:clear'
+      # end
+    end
+  end
+
+end
